Changeed property  Dictance to float

+++ WebApplication/Models/Participants.cs
@@ -11,7 +11,7 @@ namespace WebApplication.Models
         public int Id { get; set; }
         public string PartcipantName { get; set; }
         public int runnerNamber { get; set; }
-        public int Distance { get; set; }
+        public float Distance { get; set; }
         [DataType(DataType.Date), DisplayFormat(DataFormatString = "{0:H:mm}", ApplyFormatInEditMode = true)]
         public DateTime Time { get; set; }
